overide sync via async for async client
