+++ src/async_client/client.py
@@ -1,24 +1,24 @@
 __all__ = [
     'AsyncTolokaClient',
 ]
-
 import asyncio
 import datetime
 import functools
 import logging
 import threading
 from decimal import Decimal
-from typing import Optional
+from typing import Optional, Callable, List
 
 import attr
 import httpx
 
-from ..client import TolokaClient, structure, unstructure, batch_create_results
+from ..client import TolokaClient, structure, unstructure
 from ..client.exceptions import (
     raise_on_api_error,
     ValidationApiError,
 )
 from ..client.operations import Operation
+from ..client.primitives.parameter import Parameters
 from ..client.primitives.retry import AsyncRetryingOverURLLibRetry
 from ..util._managing_headers import add_headers
 from ..util.async_utils import generate_async_methods_from
@@ -107,19 +107,6 @@ class AsyncTolokaClient:
         for item in items:
             yield item
 
-    async def _collect_from_pools(self, get_method, pools):
-        items = {}
-        for pool_id, numerated_ids in pools.items():
-            obj_it = get_method(
-                pool_id=pool_id,
-                id_gte=min(numerated_ids.keys()),
-                id_lte=max(numerated_ids.keys()),
-            )
-            async for obj in obj_it:
-                if obj.id in numerated_ids:
-                    items[numerated_ids[obj.id]] = obj
-        return items
-
     @add_headers('async_client')
     async def wait_operation(
         self,
@@ -148,27 +135,80 @@ class AsyncTolokaClient:
             if datetime.datetime.now(datetime.timezone.utc) > wait_until_time:
                 raise TimeoutError
 
-    async def _sync_via_async(
-            self, objects, parameters, url: str,
-            result_type, operation_type, output_id_field: str, get_method,
+    async def _sync_via_async_pool_related(
+            self,
+            objects,
+            parameters: Parameters,
+            url: str,
+            result_type,
+            operation_type: Operation,
+            output_id_field: str,
+            get_method: Callable,
     ):
-        # TODO: do not duplicate so many things here
         if not parameters.async_mode:
             response = await self._request('post', url, json=unstructure(objects), params=unstructure(parameters))
             return structure(response, result_type)
-
-        # Emulates synchronous operation, through asynchronous calls and reading operation logs.
-
-        # Index objects to restore sequence in the future
         is_single = not isinstance(objects, list)
-        if is_single:
-            objects._unexpected['__item_idx'] = '0'
-        else:
-            for item_idx, obj in enumerate(objects):
-                obj._unexpected['__item_idx'] = str(item_idx)
+        insert_operation = await self._start_sync_via_async(objects, parameters, url, operation_type)
 
-        insert_operation = await self._async_create_objects_idempotent(url, objects, parameters, operation_type)
-        insert_operation = await self.wait_operation(insert_operation, datetime.timedelta(minutes=60))
+        pools = {}
+        validation_errors = {}
+        for log_item in await self.get_operation_log(insert_operation.id):
+            if '__item_idx' in log_item.input:
+                index = log_item.input['__item_idx']
+            else:
+                continue  # operation could be not just creating objects (e.g. open_pool while creating_object)
+            if log_item.success:
+                numerated_ids = pools.setdefault(log_item.input['pool_id'], {})
+                numerated_ids[log_item.output[output_id_field]] = index
+            else:
+                validation_errors[index] = log_item.output
+
+        # Like as in sync methods Exception will raise
+        # even if the skip_invalid_items=True but no objects are created
+        if validation_errors and not pools:
+            raise ValidationApiError(
+                code='VALIDATION_ERROR',
+                message='Validation failed',
+                payload=validation_errors,
+            )
+
+        if is_single:
+            pool_id = list(pools.keys())[0]
+            item_id = list(pools[pool_id].keys())[0]
+            return await get_method(item_id)
+        else:
+            items = await self._collect_from_pools(get_method, pools)
+            return result_type(items=items, validation_errors=validation_errors or {})
+
+    async def _collect_from_pools(self, get_method, pools):
+        items = {}
+        for pool_id, numerated_ids in pools.items():
+            obj_it = get_method(
+                pool_id=pool_id,
+                id_gte=min(numerated_ids.keys()),
+                id_lte=max(numerated_ids.keys()),
+            )
+            async for obj in obj_it:
+                if obj.id in numerated_ids:
+                    items[numerated_ids[obj.id]] = obj
+        return items
+
+    async def _sync_via_async(
+            self,
+            objects: List,
+            parameters: Parameters,
+            url: str,
+            result_type,
+            operation_type: Operation,
+            output_id_field: str,
+            get_method: Callable,
+    ):
+        if not parameters.async_mode:
+            response = await self._request('post', url, json=unstructure(objects), params=unstructure(parameters))
+            return structure(response, result_type)
+        is_single = not isinstance(objects, list)
+        insert_operation = await self._start_sync_via_async(objects, parameters, url, operation_type)
 
         item_id_to_idx = {}
         validation_errors = {}
@@ -180,14 +220,10 @@ class AsyncTolokaClient:
             if log_item.success:
                 item_id_to_idx[log_item.output[output_id_field]] = index
             else:
-                validation_errors[index] = {
-                    name: structure(error, batch_create_results.FieldValidationError)
-                    for name, error in log_item.output.items()
-                }
+                validation_errors[index] = log_item.output
 
-        # Emulates the response as in a synchronous method:
-        # it will throw an exception even if the skip_invalid_items parameter is passed
-        # but no objects are created
+        # Like as in sync methods Exception will raise
+        # even if the skip_invalid_items=True but no objects are created
         if validation_errors and not item_id_to_idx:
             raise ValidationApiError(
                 code='VALIDATION_ERROR',
